Services and Managment Interfaces Created

+++ Leilao_Produtos_Spring_Boot/src/main/java/com/example/leilao_produtos_spring_boot/models/repository/AuctionRepository.java
@@ -3,5 +3,5 @@ package com.example.leilao_produtos_spring_boot.models.repository;
 import com.example.leilao_produtos_spring_boot.models.Auction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionRepository extends JpaRepository<Auction, Long> {
+public interface AuctionRepository extends JpaRepository<Auction, Integer> {
 }
