+++ Leilao_Produtos_Spring_Boot/src/main/java/com/example/leilao_produtos_spring_boot/models/repository/BiddingRepository.java
@@ -4,6 +4,6 @@ import com.example.leilao_produtos_spring_boot.models.Bidding;
 import com.example.leilao_produtos_spring_boot.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BiddingRepository extends JpaRepository<Bidding, Long> {
+public interface BiddingRepository extends JpaRepository<Bidding, Integer> {
 
 }
